implement animation dump

+++ addon/common.h
@@ -9,7 +9,7 @@
 typedef struct options_t {
   bool importMesh;
   bool importAnimation;
-  int sampleRate;
+  double sampleRate;
 } options_t;
 
 void init_fbx_sdk (FbxManager *&_fbxMgr, FbxScene *&_fbxScene);
@@ -20,7 +20,7 @@ bool load_scene (FbxManager *_fbxMgr, FbxDocument *_fbxScene, const char *_filen
 v8::Local<v8::Object> dump_scene (FbxScene *_fbxScene, options_t *_opts);
 v8::Local<v8::Value> dump_node (FbxNode *_fbxNode);
 v8::Local<v8::Value> dump_mesh (FbxMesh *_fbxMesh);
-v8::Local<v8::Array> dump_animations (FbxScene *_fbxScene, int _sampleRate);
+v8::Local<v8::Array> dump_animations (FbxScene *_fbxScene, double _sampleRate);
 
 extern std::vector<FbxMesh *> _cachedFbxMeshes;
